Add Reserved memory for upstream u-boot.
Basically, you can either use a fully upstream u-boot, but you
have to either hit the reset button every time or you have to
nuke your SPI flash to make that work.  Or... you can use the
rockchip miniloader and an upstream u-boot; with this patch it
should no longer crash.  The crash was caused by linux not knowing
that it shouldn't write to certain portions of the memory.